Style headers in DataExplorer page, Exploration page, and ExplorationEditor page
